# api/swagger/swagger-v1-full.yaml
openapi: 3.0.1
info:
  title: API
  version: "1.0"
servers:
- url: /v1/full
...
paths:
...
  /comments/{comment_id}:
    get:
      tags:
      - comments
      description: Gets a comment by ID
      operationId: Get Comment
      parameters:
      - name: comment_id
        in: path
        description: A Comment ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/comment_response'
        "500":
          description: Server error
          content: {}
...
  /explore/best-selling:
    get:
      tags:
      - explore
      summary: Get best selling tracks and playlists with related entities
      description: Get best selling tracks and/or albums with related entities
      operationId: Get Full Best Selling
      parameters:
      - name: offset
        in: query
        description: The number of items to skip. Useful for pagination (page number
          * limit)
        schema:
          type: integer
      - name: limit
        in: query
        description: The number of items to fetch
        schema:
          type: integer
      - name: user_id
        in: query
        description: The user ID of the user making the request
        schema:
          type: string
      - name: type
        in: query
        description: The type of content to filter by
        schema:
          type: string
          default: all
          enum:
          - all
          - track
          - album
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/best_selling_full_response'
        "400":
          description: Bad request
          content: {}
        "500":
          description: Server error
          content: {}
  /notifications/{user_id}:
    get:
      tags:
      - notifications
      description: Get notifications for user ID
      operationId: Get Notifications
      parameters:
      - name: user_id
        in: path
        description: A User ID
        required: true
        schema:
          type: string
      - name: timestamp
        in: query
        description: The timestamp from which to paginate
        schema:
          type: integer
      - name: group_id
        in: query
        description: The group_id form which to paginate
        schema:
          type: string
      - name: limit
        in: query
        description: The number of notifications to return
        schema:
          type: integer
      - name: valid_types
        in: query
        description: Additional valid notification types to return
        style: form
        explode: true
        schema:
          type: array
          items:
            type: string
            enum:
            - announcement
            - follow
            - repost
            - save
            - remix
            - cosign
            - create
            - tip_receive
            - tip_send
            - challenge_reward
            - repost_of_repost
            - save_of_repost
            - tastemaker
            - reaction
            - supporter_dethroned
            - supporter_rank_up
            - supporting_rank_up
            - milestone
            - track_milestone
            - track_added_to_playlist
            - playlist_milestone
            - tier_change
            - trending
            - trending_playlist
            - trending_underground
            - usdc_purchase_buyer
            - usdc_purchase_seller
            - track_added_to_purchased_album
            - request_manager
            - approve_manager_request
            - claimable_reward
            - comment
            - comment_thread
            - comment_mention
            - comment_reaction
            - listen_streak_reminder
            - fan_remix_contest_started
            - fan_remix_contest_ended
            - fan_remix_contest_ending_soon
            - fan_remix_contest_winners_selected
            - artist_remix_contest_ended
            - artist_remix_contest_ending_soon
            - artist_remix_contest_submissions
          default: [
                "repost",
                "save",
                "follow",
                "tip_send",
                "tip_receive",
                "milestone",
                "supporter_rank_up",
                "supporting_rank_up",
                "challenge_reward",
                "tier_change",
                "create",
                "remix",
                "cosign",
                "trending",
                "supporter_dethroned",
                "announcement",
                "reaction",
                "track_added_to_playlist"
              ]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/notifications_response'
  /notifications/{user_id}/playlist_updates:
    get:
      tags:
      - notifications
      description: Get playlists the user has saved that have been updated for user
        ID
      operationId: Get Playlist Updates
      parameters:
      - name: user_id
        in: path
        description: A User ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/playlist_updates_response'
  /playlists:
    get:
      tags:
      - playlists
      description: Gets a list of playlists by ID
      operationId: Get Bulk Playlists
      parameters:
      - name: user_id
        in: query
        description: The user ID of the user making the request
        schema:
          type: string
      - name: id
        in: query
        description: The ID of the playlist(s)
        style: form
        explode: true
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/full_playlist_response'
        "400":
          description: Bad request
          content: {}
        "500":
          description: Server error
          content: {}
  /playlists/by_permalink/{handle}/{slug}:
    get:
      tags:
      - playlists
      description: Get a playlist by handle and slug
      operationId: Get Playlist By Handle and Slug
      parameters:
      - name: handle
        in: path
        description: playlist owner handle
        required: true
        schema:
          type: string
      - name: slug
        in: path
        description: playlist slug
        required: true
        schema:
          type: string
      - name: user_id
        in: query
        description: The user ID of the user making the request
        schema:
          type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/full_playlist_response'
...
  /playlists/trending:
    get:
      tags:
      - playlists
      description: Returns trending playlists for a time period
      operationId: Get Trending Playlists
      parameters:
      - name: offset
        in: query
        description: The number of items to skip. Useful for pagination (page number
          * limit)
        schema:
          type: integer
      - name: limit
        in: query
        description: The number of items to fetch
        schema:
          type: integer
      - name: user_id
        in: query
        description: The user ID of the user making the request
        schema:
          type: string
      - name: time
        in: query
        description: Calculate trending over a specified time range
        schema:
          type: string
          enum:
          - week
          - month
          - year
          - allTime
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/full_trending_playlists_response'
        "400":
          description: Bad request
          content: {}
        "500":
          description: Server error
          content: {}
...
  /playlists/{playlist_id}:
    get:
      tags:
      - playlists
      description: Get a playlist by ID
      operationId: Get Playlist
      parameters:
      - name: playlist_id
        in: path
        description: A Playlist ID
        required: true
        schema:
          type: string
      - name: user_id
        in: query
        description: The user ID of the user making the request
        schema:
          type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/full_playlist_response'
  /playlists/{playlist_id}/favorites:
    get:
      tags:
      - playlists
      description: Get users that favorited a playlist
      operationId: Get Users From Playlist Favorites
      parameters:
      - name: playlist_id
        in: path
        description: A Playlist ID
        required: true
        schema:
          type: string
      - name: offset
        in: query
        description: The number of items to skip. Useful for pagination (page number
          * limit)
        schema:
          type: integer
      - name: limit
        in: query
        description: The number of items to fetch
        schema:
          type: integer
      - name: user_id
        in: query
        description: The user ID of the user making the request
        schema:
          type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/following_response'
        "400":
          description: Bad request
          content: {}
        "500":
          description: Server error
          content: {}
  /playlists/{playlist_id}/reposts:
    get:
      tags:
      - playlists
      description: Get users that reposted a playlist
      operationId: Get Users From Playlist Reposts
      parameters:
      - name: playlist_id
        in: path
        description: A Playlist ID
        required: true
        schema:
          type: string
      - name: offset
        in: query
        description: The number of items to skip. Useful for pagination (page number
          * limit)
        schema:
          type: integer
      - name: limit
        in: query
        description: The number of items to fetch
        schema:
          type: integer
      - name: user_id
        in: query
        description: The user ID of the user making the request
        schema:
          type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/following_response'
        "400":
          description: Bad request
          content: {}
        "500":
          description: Server error
          content: {}
  /playlists/{playlist_id}/tracks:
    get:
      tags:
      - playlists
      description: Fetch tracks within a playlist.
      operationId: Get Playlist Tracks
      parameters:
      - name: playlist_id
        in: path
        description: A Playlist ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/full_playlist_tracks_response'
        "400":
          description: Bad request
          content: {}
        "500":
          description: Server error
          content: {}
...
  /tracks/{track_id}:
    get:
      tags:
      - tracks
      description: Gets a track by ID.
      operationId: Get Track
      parameters:
      - name: track_id
        in: path
        description: A Track ID
        required: true
        schema:
          type: string
      - name: user_id
        in: query
        description: The user ID of the user making the request
        schema:
          type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/full_track_response'
...
components:
  schemas:
...
    notifications_response:
      required:
      - latest_chain_block
      - latest_chain_slot_plays
      - latest_indexed_block
      - latest_indexed_slot_plays
      - signature
      - timestamp
      - version
      type: object
      properties:
        latest_chain_block:
          type: integer
        latest_indexed_block:
          type: integer
        latest_chain_slot_plays:
          type: integer
        latest_indexed_slot_plays:
          type: integer
        signature:
          type: string
        timestamp:
          type: string
        version:
          $ref: '#/components/schemas/version_metadata'
        data:
          $ref: '#/components/schemas/notifications'
...
    comment_response:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/comment'
...
    best_selling_full_response:
      required:
      - latest_chain_block
      - latest_chain_slot_plays
      - latest_indexed_block
      - latest_indexed_slot_plays
      - signature
      - timestamp
      - version
      type: object
      properties:
        latest_chain_block:
          type: integer
        latest_indexed_block:
          type: integer
        latest_chain_slot_plays:
          type: integer
        latest_indexed_slot_plays:
          type: integer
        signature:
          type: string
        timestamp:
          type: string
        version:
          $ref: '#/components/schemas/version_metadata'
        data:
          type: array
          items:
            $ref: '#/components/schemas/best_selling_item'
        related:
          $ref: '#/components/schemas/related'
...